+++ sdk/webpubsub/azure-messaging-webpubsubservice/azure/messaging/webpubsubservice/__init__.py
@@ -17,7 +17,7 @@ try:
     patch_sdk()
 
     from ._patch import build_authentication_token
-    from ._patch import WebPubSubServicePatchClient as WebPubSubServiceClient
+    from ._patch import WebPubSubServiceClient
     __all__ = ['WebPubSubServiceClient', 'build_authentication_token']
 except ImportError:
     pass