+++ sdk/webpubsub/azure-messaging-webpubsubservice/azure/messaging/webpubsubservice/aio/__init__.py
@@ -13,7 +13,7 @@ try:
     from ._patch import patch_sdk  # type: ignore
     patch_sdk()
 
-    from ._patch import AsyncWebPubSubServicePatchClient as WebPubSubServiceClient
+    from ._patch import WebPubSubServiceClient
     __all__ = ['WebPubSubServiceClient']
 except ImportError:
     pass