+++ sdk/webpubsub/azure-messaging-webpubsubservice/tests/test_smoke.py
@@ -5,7 +5,7 @@
 # license information.
 # -------------------------------------------------------------------------
 from testcase import WebpubsubTest, WebpubsubPowerShellPreparer
-
+from azure.messaging.webpubsubservice.operations._operations import build_web_pub_sub_send_to_all_request
 
 class WebpubsubSmokeTest(WebpubsubTest):
 
@@ -20,6 +20,13 @@ class WebpubsubSmokeTest(WebpubsubTest):
         client.web_pub_sub.send_to_all('Hub', {'hello': 'test_webpubsub_send_to_all'})
 
     @WebpubsubPowerShellPreparer()
-    def test_webpubsub_send_to_all_api_management_proxy(self, webpubsub_endpoint, reverse_proxy_endpoint=None):
-        client = self.create_client(endpoint=webpubsub_endpoint, reverse_proxy_endpoint=reverse_proxy_endpoint)
+    def test_webpubsub_send_to_all_api_management_proxy(self, webpubsub_endpoint, webpubsub_reverse_proxy_endpoint=None):
+        client = self.create_client(endpoint=webpubsub_endpoint, reverse_proxy_endpoint=webpubsub_reverse_proxy_endpoint)
         client.web_pub_sub.send_to_all('Hub', {'hello': 'test_webpubsub_send_to_all_api_management_proxy'})
+
+    @WebpubsubPowerShellPreparer()
+    def test_webpubsub_send_request(self, webpubsub_endpoint):
+        client = self.create_client(endpoint=webpubsub_endpoint)
+        request = build_web_pub_sub_send_to_all_request('Hub', content='test_webpubsub_send_request', content_type='text/plain')
+        response = client.send_request(request)
+        assert response.status_code == 202