+++ sdk/webpubsub/azure-messaging-webpubsubservice/tests/test_smoke_async.py
@@ -21,6 +21,6 @@ class WebpubsubSmokeTestAsync(WebpubsubTestAsync):
         await client.web_pub_sub.send_to_all('Hub', {'hello': 'test_webpubsub_send_to_all'})
 
     @WebpubsubPowerShellPreparer()
-    async def test_webpubsub_send_to_all_apim_proxy(self, webpubsub_endpoint, reverse_proxy_endpoint=None):
-        client = self.create_client(endpoint=webpubsub_endpoint, reverse_proxy_endpoint=reverse_proxy_endpoint)
+    async def test_webpubsub_send_to_all_apim_proxy(self, webpubsub_endpoint, webpubsub_reverse_proxy_endpoint=None):
+        client = self.create_client(endpoint=webpubsub_endpoint, reverse_proxy_endpoint=webpubsub_reverse_proxy_endpoint)
         await client.web_pub_sub.send_to_all('Hub', {'hello': 'test_webpubsub_send_to_all_apim_proxy'})