+++ sdk/webpubsub/azure-messaging-webpubsubservice/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add operations to client
 - Support AAD
+- Support Api Management Proxy
 
 ## 1.0.0b1 (2021-04-27)