+++ sdk/webpubsub/azure-messaging-webpubsubservice/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Release History
 
-## 1.0.0b2 (Unreleased)
+## 1.0.0b2 (2021-10-13)
 
+- Change api-version to `2021-10-01`  
 - Add operations to client
 - Support AAD
 - Support Api Management Proxy